requirements.txt
discord.py>=2.3.2
yt-dlp>=2024.1.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
spotipy>=2.23.0